+++ Runtime/Api/GameProfilesApi.cs
@@ -11,7 +11,7 @@ using System.Collections.Generic;
 
 namespace AccelByte.Api
 {
-    [Obsolete("This interface is deprecated and will be removed on 2025.5.AGS. Please use Api.CloudSave instead")]
+    [Obsolete("This interface is deprecated. Please use Api.CloudSave instead")]
     public class GameProfilesApi : ApiBase
     {
         /// <summary>